Change stock type from Int to Double

+++ app/src/main/java/com/grocerypos/v11/PartyDashboardActivity.kt
@@ -135,7 +135,7 @@ class PartyDashboardActivity : AppCompatActivity() {
         val cost: Double,
         val salePrice: Double,
         val wholesalePrice: Double,
-        val stock: Int,
+        val stock: Double,
         val stockDisplay: String,
         val soldQty: Int,
         val soldAmt: Double,
